Improve and fix #542718 – Creating large ISO fails due to filesystem restrictions 	Also fix variable overflow in mkisofs, genisoimage, cdrdao when they reported 	size of output
	* src/brasero-drive-properties.c
	(brasero_drive_properties_set_tmpdir_info),
	(brasero_drive_properties_tmpdir_changed_cb),
	(brasero_drive_properties_set_tmpdir),
	(brasero_drive_properties_init),
	(brasero_drive_properties_finalize):
	* src/burn-job.c (brasero_job_check_output_volume_space):
	* src/plugins/cdrdao/burn-cdrdao.c
	(brasero_cdrdao_read_stderr_image), (brasero_cdrdao_set_argv):
	* src/plugins/cdrkit/burn-genisoimage.c
	(brasero_genisoimage_read_isosize):
	* src/plugins/cdrkit/burn-readom.c (brasero_readom_get_size):
	* src/plugins/cdrtools/burn-mkisofs.c
	(brasero_mkisofs_read_isosize):
	* src/plugins/cdrtools/burn-readcd.c (brasero_readcd_get_size):
	* src/plugins/dvdcss/burn-dvdcss-private.h:
	* src/plugins/growisofs/burn-growisofs.c
	(brasero_growisofs_read_stderr):

svn path=/trunk/; revision=1024

+++ src/plugins/cdrdao/burn-cdrdao.c
@@ -70,7 +70,7 @@ brasero_cdrdao_read_stderr_image (BraseroCdrdao *cdrdao, const gchar *line)
 		brasero_job_get_action (BRASERO_JOB (cdrdao), &action);
 		if (action == BRASERO_JOB_ACTION_SIZE) {
 			/* get the number of sectors. As we added -raw sector = 2352 bytes */
-			brasero_job_set_output_size_for_current_track (BRASERO_JOB (cdrdao), s1, s1 * 2352);
+			brasero_job_set_output_size_for_current_track (BRASERO_JOB (cdrdao), s1, (gint64) s1 * 2352ULL);
 			brasero_job_finished_session (BRASERO_JOB (cdrdao));
 		}
 	}
@@ -495,7 +495,7 @@ brasero_cdrdao_set_argv (BraseroProcess *process,
 
 			brasero_job_set_output_size_for_current_track (BRASERO_JOB (cdrdao),
 								       sectors,
-								       sectors * 2352);
+								       sectors * 2352ULL);
 		}
 		else
 			return BRASERO_BURN_NOT_SUPPORTED;
